Prezi.com
Ready to ship.

One minor issue that brings up a possible .json enhancement:
blog.prezi.com displays the Waltz button even though it is not a site
for users to log in. Do we need a new parameter that allows us to
exclude/ignore specific URLs (or patterns of URLs)?